requirements.txt
streamlit
crewai
groq
python-dotenv
pandas
matplotlib
seaborn